Point & Threshold Operation
Point operation is changing the intensity value of a single pixel. This operation can be done to brighten up or darken the image.
Threshold Operation is an operation for classifying the pixel values into groups separated by a threshold number; if a pixel value is below the threshold value, that pixel value will be categorized into ‘zero’. Pixel value that has a higher value than the threshold will be categorized into ‘one’. From this, it can be called a binary image.