Create Filter
A filter is a function used for blurring or sharpening the image. It is similar to the point operation but the filter will take neighbouring pixels into account and the weight of each pixel is assigned also. There are 2 types of filter which are the linear and the non-linear filter. 
For this file, the linear filter is focused. Linear filters can be categorized into 3 methods which are the box filter, Gaussian filter, Mexican hat filter.

The simplest method is the box filter where each pixel will have the same weight which is 1. The way to find the intensity value of the focused pixel is by finding the average intensity values from its intensity value and its neighbouring pixels’. Adding a mathematical model into account, the gaussian distribution is used to find the weight of each pixel value. Lastly, the Mexican hat filter is similar to the Gaussian filter but the intensity values can be negative integers. Thus, sharpening the image. The function used for finding the weight is as shown below.