Non Linear Filter
- Separable 2D Filters - 
By adding filters to an image, it would take quite some time for the code to run through all the image pixels. Doing Box Filter with the kernel matrix of 3x3 for example, the code will have to run and calculate values for 9 times in order to have all of the values used for calculating a new pixel value. The way to shorten the time is by separating the kernel matrix and then do a convolution to the image. By this, the code will have to run for 6 times in order to have all of the values used for calculating a new pixel value.
1. MAX MIN MEDIAN FILTERS
2. WEIGHT MEDIAN FILTERS