Create EdgeDetection
Edge detection is a way to determine the edge of a picture to gain contrast and make the edge clearer. The key to edge detection is doing a convolution in x- and y-axes with the original picture and then using the formula ‘sqrt(x2 + y2)’ to get the new pixel value.
There are 3 main filters which are the Prewitt, Sobel, and Roberts filters.