Create MorphologicalFilters
The morphological filter is a type of image processing operation for shrinking or dilating the image. The binary image [an image that consists of pixel values of 1 (white) and 0 (black)] is mostly used with this type of operation. There are 2 main types which are erosion and dilation. As the name suggested, erosion is a filter for shrinking or thinning the image. As opposed to dilation which is a filter for enlarging the image.

The combination of the 2 filters can help smoothes the contour object. For opening (erosion before doing dilation), it helps smoothes the object and eliminate the thin protrusions. For closing (dilation before doing erosion), helps smoothes the object, fuses narrow brakes, and fill gaps or small holes.